+++ src/libs/zbxcrypto/base64.c
@@ -216,7 +216,7 @@ void	str_base64_encode_dyn(const char *p_str, char **p_b64str, int in_size)
  *             p_out_size - [OUT] the size (length) of the str decoded        *
  *                                                                            *
  ******************************************************************************/
-void	str_base64_decode(const char *p_b64str, char *p_str, int maxsize, int *p_out_size)
+void	str_base64_decode(const char *p_b64str, char *p_str, size_t maxsize, size_t *p_out_size)
 {
 	const char	*p;
 	char		from[4];
